+++ src/main/java/be/woutschoovaerts/mollie/data/profile/ProfileResponse.java
@@ -30,6 +30,8 @@ public class ProfileResponse {
 
     private int categoryCode;
 
+    private BusinessCategory businessCategory;
+
     private ProfileStatus status;
 
     private ProfileReviewResponse review;