+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## 3.7.2
  - Added payment link support
+ - Mark 'categoryCode' parameter in Profile API as deprecated
+ - Added 'businessCategory' as parameter in Profile API together with a new BusinessCategory enumeration
+
 
 ## 3.7.1
  - Make chargeback reason a working Optional instead of null.